requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
gunicorn
attrs==18.2.0
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.0.3
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
networkx==2.5.1
numpy==1.20.3
pandas==1.2.4
pdflatex==0.1.3
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
tex==1.8
tifffile==2021.4.8
torch==1.8.1+cpu 
torchvision==0.9.1+cpu
typing-extensions==3.10.0.0
Werkzeug==1.0.1
